requirements.txt
backports.csv==1.0.7
defusedxml==0.5.0
diff-match-patch==20181111
Django==2.1.7
django-common-helpers==0.9.2
django-cron==0.5.1
django-filter==2.2.0
django-import-export==1.2.0
djangorestframework==3.9.2
et-xmlfile==1.0.1
jdcal==1.4
numpy==1.16.2
odfpy==1.4.0
openpyxl==2.6.1
pandas==0.24.2
pkg-resources==0.0.0
python-dateutil==2.8.0
pytz==2018.9
PyYAML==5.1
six==1.12.0
tablib==0.13.0
xlrd==1.2.0
xlwt==1.3.0
